+++ CHANGES.rst
@@ -11,6 +11,8 @@ releases are available on `Anaconda.org
 ------------------
 
 - :gh:`2` replaces the pre-commit pipeline step with pre-commit.ci.
+- :gh:`11` implement :func:`pybaum.tree_util.tree_yield` and
+  :func:`pybaum.tree_util.tree_just_yield`.
 
 
 0.1.0 - 2022-01-28